Time series prediction using Neural network
An experimental learning on time series- Mackay Glass, which is generated via the formula and then split into training and validation sets. The first 500 are used for training an Artificial Neural Network. The resulting model is used to predict the next 500 data points. As can be seen, the predicted points are far more uniformly distributed as compared to the generated set